requirements.txt
qiskit
qiskit-aer
pycryptodome
matplotlib
numpy